Refactor: 폴더 및 파일 구조 정리
