+++ src/app/nearby-stores/nearby-stores-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core'
 import { Routes, RouterModule } from '@angular/router'
 
-import { NearbyStoresPage } from './nearby-stores.page'
+import { NearbyStoresPage } from './pages/nearby-stores.page'
 
 
 const routes: Routes = [
