+++ src/app/nearby-stores/nearby-stores.module.ts
@@ -5,8 +5,8 @@ import { FormsModule } from '@angular/forms'
 import { IonicModule } from '@ionic/angular'
 
 import { NearbyStoresPageRoutingModule } from './nearby-stores-routing.module'
-import { NearbyStoresPage } from './nearby-stores.page'
-import { NearbyStoresComponent } from '../components/nearby-stores.component'
+import { NearbyStoresPage } from './pages/nearby-stores.page'
+import { NearbyStoresComponent } from './components/nearby-stores.component'
 import { CommonComponentsModule } from 'src/app/model/common/common.module'
 
 @NgModule({
